+++ core/src/blocks/mod.rs
@@ -1,19 +1,24 @@
-use scale_decode::DecodeAsFields;
+// Copyright 2019-2024 Parity Technologies (UK) Ltd.
+// This file is dual-licensed as Apache-2.0 or GPL-3.0.
+// see LICENSE for license details.
 
-/// Trait to uniquely identify the extrinsic's identity from the runtime metadata.
-///
-/// Generated API structures that represent an extrinsic implement this trait.
-///
-/// The trait is utilized to decode emitted extrinsics from a block, via obtaining the
-/// form of the `Extrinsic` from the metadata.
-pub trait StaticExtrinsic: DecodeAsFields {
-    /// Pallet name.
-    const PALLET: &'static str;
-    /// Call name.
-    const CALL: &'static str;
+//! Functions and types for decoding and iterating over block bodies.
 
-    /// Returns true if the given pallet and call names match this extrinsic.
-    fn is_extrinsic(pallet: &str, call: &str) -> bool {
-        Self::PALLET == pallet && Self::CALL == call
-    }
-}
+mod extrinsics;
+mod extrinsic_signed_extensions;
+mod static_extrinsic;
+
+pub use static_extrinsic::StaticExtrinsic;
+
+pub use extrinsic_signed_extensions::{
+    ExtrinsicSignedExtensions,
+    ExtrinsicSignedExtension,
+};
+
+pub use extrinsics::{
+    Extrinsics,
+    ExtrinsicDetails,
+    SignedExtrinsicDetails,
+    FoundExtrinsic,
+    ExtrinsicMetadataDetails,
+};
