+++ core/src/custom_values/custom_value_address.rs
@@ -1,3 +1,7 @@
+// Copyright 2019-2024 Parity Technologies (UK) Ltd.
+// This file is dual-licensed as Apache-2.0 or GPL-3.0.
+// see LICENSE for license details.
+
 use derive_where::derive_where;
 
 use crate::dynamic::DecodedValueThunk;
