+++ core/src/events.rs
@@ -1,3 +1,9 @@
+// Copyright 2019-2024 Parity Technologies (UK) Ltd.
+// This file is dual-licensed as Apache-2.0 or GPL-3.0.
+// see LICENSE for license details.
+
+//! Types and functions for decoding and iterating over events.
+
 use alloc::sync::Arc;
 use alloc::vec::Vec;
 use codec::{Compact, Decode, Encode};
@@ -276,12 +282,12 @@ impl<T: Config> EventDetails<T> {
 
     /// The name of the pallet from whence the Event originated.
     pub fn pallet_name(&self) -> &str {
-        self.event_metadata().pallet().name()
+        self.event_metadata().pallet.name()
     }
 
     /// The name of the event (ie the name of the variant that it corresponds to).
     pub fn variant_name(&self) -> &str {
-        &self.event_metadata().variant().name
+        &self.event_metadata().variant.name
     }
 
     /// Fetch details from the metadata for this event.
@@ -370,17 +376,10 @@ impl<T: Config> EventDetails<T> {
 
 /// Details for the given event plucked from the metadata.
 pub struct EventMetadataDetails<'a> {
-    pallet: PalletMetadata<'a>,
-    variant: &'a scale_info::Variant<scale_info::form::PortableForm>,
-}
-
-impl<'a> EventMetadataDetails<'a> {
-    pub fn pallet(&self) -> PalletMetadata<'a> {
-        self.pallet
-    }
-    pub fn variant(&self) -> &'a scale_info::Variant<scale_info::form::PortableForm> {
-        self.variant
-    }
+    /// Metadata for the pallet that the event belongs to.
+    pub pallet: PalletMetadata<'a>,
+    /// Metadata for the variant which describes the pallet events.
+    pub variant: &'a scale_info::Variant<scale_info::form::PortableForm>,
 }
 
 /// Event related test utilities used outside this module.
@@ -504,8 +503,9 @@ pub(crate) mod test_utils {
             },
         );
         let runtime_metadata: RuntimeMetadataPrefixed = meta.into();
+        let metadata: subxt_metadata::Metadata = runtime_metadata.try_into().unwrap();
 
-        Metadata::new(runtime_metadata.try_into().unwrap())
+        Metadata::from(metadata)
     }
 
     /// Build an `Events` object for test purposes, based on the details provided,
