+++ core/src/macros.rs
@@ -1,3 +1,7 @@
+// Copyright 2019-2024 Parity Technologies (UK) Ltd.
+// This file is dual-licensed as Apache-2.0 or GPL-3.0.
+// see LICENSE for license details.
+
 macro_rules! cfg_feature {
 	($feature:literal, $($item:item)*) => {
 		$(
