+++ core/src/metadata/metadata_type.rs
@@ -1,4 +1,4 @@
-// Copyright 2019-2023 Parity Technologies (UK) Ltd.
+// Copyright 2019-2024 Parity Technologies (UK) Ltd.
 // This file is dual-licensed as Apache-2.0 or GPL-3.0.
 // see LICENSE for license details.
 
@@ -21,12 +21,6 @@ impl core::ops::Deref for Metadata {
 }
 
 impl Metadata {
-    pub fn new(md: subxt_metadata::Metadata) -> Self {
-        Metadata {
-            inner: Arc::new(md),
-        }
-    }
-
     /// Identical to `metadata.pallet_by_name()`, but returns an error if the pallet is not found.
     pub fn pallet_by_name_err(
         &self,
@@ -53,11 +47,21 @@ impl Metadata {
         self.runtime_api_trait_by_name(name)
             .ok_or_else(|| MetadataError::RuntimeTraitNotFound(name.to_owned()))
     }
+
+    /// Identical to `metadata.custom().get(name)`, but returns an error if the trait is not found.
+    pub fn custom_value_by_name_err(
+        &self,
+        name: &str,
+    ) -> Result<subxt_metadata::CustomValueMetadata, MetadataError> {
+        self.custom()
+            .get(name)
+            .ok_or_else(|| MetadataError::CustomValueNameNotFound(name.to_owned()))
+    }
 }
 
 impl From<subxt_metadata::Metadata> for Metadata {
     fn from(md: subxt_metadata::Metadata) -> Self {
-        Metadata::new(md)
+        Metadata { inner: Arc::new(md) }
     }
 }
 
@@ -70,68 +74,68 @@ impl TryFrom<frame_metadata::RuntimeMetadataPrefixed> for Metadata {
 
 impl codec::Decode for Metadata {
     fn decode<I: codec::Input>(input: &mut I) -> Result<Self, codec::Error> {
-        subxt_metadata::Metadata::decode(input).map(Metadata::new)
+        subxt_metadata::Metadata::decode(input).map(Metadata::from)
     }
 }
 
-/// Some extension methods on [`subxt_metadata::Metadata`] that return Errors instead of Options.
-pub trait MetadataExt {
-    fn pallet_by_name_err(
-        &self,
-        name: &str,
-    ) -> Result<subxt_metadata::PalletMetadata, MetadataError>;
+// /// Some extension methods on [`subxt_metadata::Metadata`] that return Errors instead of Options.
+// pub trait MetadataExt {
+//     fn pallet_by_name_err(
+//         &self,
+//         name: &str,
+//     ) -> Result<subxt_metadata::PalletMetadata, MetadataError>;
 
-    fn pallet_by_index_err(
-        &self,
-        index: u8,
-    ) -> Result<subxt_metadata::PalletMetadata, MetadataError>;
+//     fn pallet_by_index_err(
+//         &self,
+//         index: u8,
+//     ) -> Result<subxt_metadata::PalletMetadata, MetadataError>;
 
-    fn runtime_api_trait_by_name_err(
-        &self,
-        name: &str,
-    ) -> Result<subxt_metadata::RuntimeApiMetadata, MetadataError>;
+//     fn runtime_api_trait_by_name_err(
+//         &self,
+//         name: &str,
+//     ) -> Result<subxt_metadata::RuntimeApiMetadata, MetadataError>;
 
-    fn custom_value_by_name_err(
-        &self,
-        name: &str,
-    ) -> Result<subxt_metadata::CustomValueMetadata, MetadataError>;
-}
+//     fn custom_value_by_name_err(
+//         &self,
+//         name: &str,
+//     ) -> Result<subxt_metadata::CustomValueMetadata, MetadataError>;
+// }
 
-impl MetadataExt for subxt_metadata::Metadata {
-    /// Identical to `metadata.pallet_by_name()`, but returns an error if the pallet is not found.
-    fn pallet_by_name_err(
-        &self,
-        name: &str,
-    ) -> Result<subxt_metadata::PalletMetadata, MetadataError> {
-        self.pallet_by_name(name)
-            .ok_or_else(|| MetadataError::PalletNameNotFound(name.to_owned()))
-    }
+// impl MetadataExt for subxt_metadata::Metadata {
+//     /// Identical to `metadata.pallet_by_name()`, but returns an error if the pallet is not found.
+//     fn pallet_by_name_err(
+//         &self,
+//         name: &str,
+//     ) -> Result<subxt_metadata::PalletMetadata, MetadataError> {
+//         self.pallet_by_name(name)
+//             .ok_or_else(|| MetadataError::PalletNameNotFound(name.to_owned()))
+//     }
 
-    /// Identical to `metadata.pallet_by_index()`, but returns an error if the pallet is not found.
-    fn pallet_by_index_err(
-        &self,
-        index: u8,
-    ) -> Result<subxt_metadata::PalletMetadata, MetadataError> {
-        self.pallet_by_index(index)
-            .ok_or(MetadataError::PalletIndexNotFound(index))
-    }
+//     /// Identical to `metadata.pallet_by_index()`, but returns an error if the pallet is not found.
+//     fn pallet_by_index_err(
+//         &self,
+//         index: u8,
+//     ) -> Result<subxt_metadata::PalletMetadata, MetadataError> {
+//         self.pallet_by_index(index)
+//             .ok_or(MetadataError::PalletIndexNotFound(index))
+//     }
 
-    /// Identical to `metadata.runtime_api_trait_by_name()`, but returns an error if the trait is not found.
-    fn runtime_api_trait_by_name_err(
-        &self,
-        name: &str,
-    ) -> Result<subxt_metadata::RuntimeApiMetadata, MetadataError> {
-        self.runtime_api_trait_by_name(name)
-            .ok_or_else(|| MetadataError::RuntimeTraitNotFound(name.to_owned()))
-    }
+//     /// Identical to `metadata.runtime_api_trait_by_name()`, but returns an error if the trait is not found.
+//     fn runtime_api_trait_by_name_err(
+//         &self,
+//         name: &str,
+//     ) -> Result<subxt_metadata::RuntimeApiMetadata, MetadataError> {
+//         self.runtime_api_trait_by_name(name)
+//             .ok_or_else(|| MetadataError::RuntimeTraitNotFound(name.to_owned()))
+//     }
 
-    /// Identical to `metadata.runtime_api_trait_by_name()`, but returns an error if the trait is not found.
-    fn custom_value_by_name_err(
-        &self,
-        name: &str,
-    ) -> Result<subxt_metadata::CustomValueMetadata, MetadataError> {
-        self.custom()
-            .get(name)
-            .ok_or_else(|| MetadataError::CustomValueNameNotFound(name.to_owned()))
-    }
-}
+//     /// Identical to `metadata.runtime_api_trait_by_name()`, but returns an error if the trait is not found.
+//     fn custom_value_by_name_err(
+//         &self,
+//         name: &str,
+//     ) -> Result<subxt_metadata::CustomValueMetadata, MetadataError> {
+//         self.custom()
+//             .get(name)
+//             .ok_or_else(|| MetadataError::CustomValueNameNotFound(name.to_owned()))
+//     }
+// }
