+++ core/src/metadata/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2019-2023 Parity Technologies (UK) Ltd.
+// Copyright 2019-2024 Parity Technologies (UK) Ltd.
 // This file is dual-licensed as Apache-2.0 or GPL-3.0.
 // see LICENSE for license details.
 
@@ -8,7 +8,4 @@ mod decode_encode_traits;
 mod metadata_type;
 
 pub use decode_encode_traits::{DecodeWithMetadata, EncodeWithMetadata};
-pub use metadata_type::{Metadata, MetadataExt};
-
-// Expose metadata types under a sub module in case somebody needs to reference them:
-pub use subxt_metadata as types;
+pub use metadata_type::Metadata;
