+++ core/src/runtime_api.rs
@@ -1,6 +1,8 @@
-// Copyright 2019-2023 Parity Technologies (UK) Ltd.
+// Copyright 2019-2024 Parity Technologies (UK) Ltd.
 // This file is dual-licensed as Apache-2.0 or GPL-3.0.
 // see LICENSE for license details.
+
+//! Types and functions for constructing runtime API requests.
 
 use alloc::borrow::Cow;
 use alloc::borrow::ToOwned;
