+++ core/src/storage/storage_key.rs
@@ -1,3 +1,7 @@
+// Copyright 2019-2024 Parity Technologies (UK) Ltd.
+// This file is dual-licensed as Apache-2.0 or GPL-3.0.
+// see LICENSE for license details.
+
 use super::utils::hash_bytes;
 use crate::{
     error::{Error, MetadataError, StorageAddressError},
