+++ subxt/examples/setup_client_offline.rs
@@ -16,7 +16,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     };
 
     // 2. A runtime version (system_version constant on a Substrate node has these):
-    let runtime_version = subxt::client::RuntimeVersion::new(9370, 20);
+    let runtime_version = subxt::client::RuntimeVersion {
+        spec_version: 9370,
+        transaction_version: 20
+    };
 
     // 3. Metadata (I'll load it from the downloaded metadata, but you can use
     //    `subxt metadata > file.scale` to download it):
