+++ subxt/src/backend/legacy/mod.rs
@@ -181,16 +181,16 @@ impl<T: Config + Send + Sync + 'static> Backend<T> for LegacyBackend<T> {
 
     async fn current_runtime_version(&self) -> Result<RuntimeVersion, Error> {
         let details = self.methods.state_get_runtime_version(None).await?;
-        Ok(RuntimeVersion::new(
-            details.spec_version,
-            details.transaction_version,
-        ))
+        Ok(RuntimeVersion {
+            spec_version: details.spec_version,
+            transaction_version: details.transaction_version,
+        })
     }
 
     async fn stream_runtime_version(&self) -> Result<StreamOfResults<RuntimeVersion>, Error> {
         let sub = self.methods.state_subscribe_runtime_version().await?;
         let sub =
-            sub.map(|r| r.map(|v| RuntimeVersion::new(v.spec_version, v.transaction_version)));
+            sub.map(|r| r.map(|v| RuntimeVersion { spec_version: v.spec_version, transaction_version: v.transaction_version }));
         Ok(StreamOf(Box::pin(sub)))
     }
